requirements.txt
asttokens==2.2.1
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
decorator==5.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
executing==1.2.0
Flask==2.2.3
Flask-Cors==3.0.10
idna==3.4
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
mysql-connector-python==8.0.32
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
preshed==3.0.8
prompt-toolkit==3.0.37
protobuf==3.20.3
pure-eval==0.2.2
pydantic==1.10.5
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
skillNer==1.0.3
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
thinc==8.1.7
threadpoolctl==3.1.0
tqdm==4.64.1
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.2.3
